+++ src/types.ts
@@ -28,6 +28,7 @@ export type ALTestAssembly = {
 		'run-date': string;
 		'test-framework': string;
 		name: string;
+		errors?: string;
 	};
 	collection: ALTestCollection[];
 };
@@ -148,7 +149,8 @@ export enum enableCodeCoverage {
 }
 
 export type invokeTestRunnerOptions = {
-	enableCodeCoverage: boolean
+	enableCodeCoverage: boolean;
+	onOutput?: (data: string) => void;
 }
 
 export type DevOpsTestStep = {